+++ lib/src/utils/ds_message_content_type.util.dart
@@ -2,4 +2,5 @@ abstract class DSMessageContentType {
   static const String textPlain = 'text/plain';
   static const String sensitive = 'application/vnd.lime.sensitive+json';
   static const String mediaLink = 'application/vnd.lime.media-link+json';
+  static const String select = 'application/vnd.lime.select+json';
 }