+++ lib/src/utils/ds_utils.util.dart
@@ -2,4 +2,6 @@
 abstract class DSUtils {
   static const packageName = 'blip_ds';
   static const defaultAnimationDuration = Duration(milliseconds: 300);
+  static const double bubbleMinSize = 240.0;
+  static const double bubbleMaxSize = 480.0;
 }