+++ lib/src/widgets/chat/ds_message_bubble.widget.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import '../../enums/ds_align.enum.dart';
 import '../../enums/ds_border_radius.enum.dart';
 import '../../models/ds_message_bubble_style.model.dart';
+import '../../utils/ds_utils.util.dart';
 import '../animations/ds_animated_size.widget.dart';
 
 class DSMessageBubble extends StatelessWidget {
@@ -10,6 +11,9 @@ class DSMessageBubble extends StatelessWidget {
   final Widget child;
   final List<DSBorderRadius> borderRadius;
   final EdgeInsets padding;
+  final bool shouldUseDefaultSize;
+  final double defaultMaxSize;
+  final double defaultMinSize;
   final DSMessageBubbleStyle style;
 
   const DSMessageBubble({
@@ -21,6 +25,9 @@ class DSMessageBubble extends StatelessWidget {
       vertical: 8.0,
       horizontal: 16.0,
     ),
+    this.shouldUseDefaultSize = false,
+    this.defaultMaxSize = DSUtils.bubbleMaxSize,
+    this.defaultMinSize = DSUtils.bubbleMinSize,
     required this.style,
   }) : super(key: key);
 
@@ -36,14 +43,20 @@ class DSMessageBubble extends StatelessWidget {
       flex: 5,
       child: DSAnimatedSize(
         child: Container(
+          padding: const EdgeInsets.all(1.0),
           decoration: BoxDecoration(
             borderRadius: _getBorderRadius(),
             color: style.bubbleBorderColor(align),
           ),
-          padding: const EdgeInsets.all(1.0),
           child: ClipRRect(
             borderRadius: _getBorderRadius(),
             child: Container(
+              constraints: shouldUseDefaultSize
+                  ? BoxConstraints(
+                      maxWidth: defaultMaxSize,
+                      minWidth: defaultMinSize,
+                    )
+                  : null,
               padding: padding,
               color: style.bubbleBackgroundColor(align),
               child: child,