+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.14
+
+* Added suport for Quick Reply and Select
+* Added suport for Video messages
+* Image message bubble size fixe
+
 ## 0.0.13
 
 * Added delivery report & timestamp message widget